Refactor Sidebar into multiple files
